Use same data as in mapbox-style example
