+++ examples/offscreen-canvas-tiles.css
@@ -1,3 +1,3 @@
 .map {
-  background: #f8f4f0;
+  background: rgba(232, 230, 223, 1);
 }
